requirements.txt
# --- Core & Orchestration ---
langchain
langchain-community
langgraph
python-dotenv

# --- LLM Providers ---
langchain-mistralai
langchain-openai

# --- RAG & Vector Stores ---
langchain-chroma
chromadb
langchain-huggingface
sentence-transformers

# --- Document Processing ---
pypdf
markdown
xhtml2pdf

# --- Tools & Search ---
tavily-python

# --- Monitoring & Data ---
mlflow
pandas
spacy
